First Commit
First commit of the app to record and play videos.

Can also edit the video via trimming using the UIVideoEditorController.
Right now, it’s not quite implemented well — displays after user
records a video in a two-step process.

Next, may want to display the ability to trim before dismissing the
UIImagePickerController and the UIAlertView shows.